Added check if file path to be deleted is previous version.

+++ testnet/register-node.js
@@ -102,7 +102,7 @@ function checkForUpdate() {
         execSync(`cp -af ${appMigrationDirPath}/. ${configDir}`);
 
         // Potential risk of race condition here. Coping and linking has to be atomic operation.
-
+        const previousVersionPath = fs.realpathSync('/ot-node/current');
         // Just replace current link.
         execSync(`ln -fns ${updateInfo.path} /ot-node/current`);
 
@@ -111,7 +111,8 @@ function checkForUpdate() {
             const filePath = `/ot-node/${fileName}`;
             if (fs.lstatSync(filePath).isDirectory()
                 && filePath !== updateInfo.path
-                 && /^\d+\.\d+\.\d+$/.test(fileName)) {
+                && filePath !== previousVersionPath
+                && /^\d+\.\d+\.\d+$/.test(fileName)) {
                 fs.rmdirSync(filePath);
                 logger.trace(`Successfully removed old version directory: ${filePath}`);
             }
